fix(server): second shard tests
