+++ packages/twenty-server/src/engine/core-modules/view/flat-view/constants/flat-view-properies-to-compare.constant.ts
@@ -3,5 +3,6 @@ import { type FlatView } from 'src/engine/core-modules/view/flat-view/types/flat
 
 export const FLAT_VIEW_PROPERTIES_TO_COMPARE = [
   'key',
+  'deletedAt',
   ...FLAT_VIEW_EDITABLE_PROPERTIES,
 ] as const satisfies (keyof FlatView)[];
