+++ packages/twenty-server/src/engine/core-modules/view/services/view-v2.service.ts
@@ -281,17 +281,6 @@ export class ViewV2Service {
       );
     }
 
-    const { flatViewMaps: recomputedExistingFlatViewMaps } =
-      await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
-        {
-          workspaceId,
-          flatEntities: ['flatViewMaps'],
-        },
-      );
-
-    return findFlatEntityByIdInFlatEntityMapsOrThrow({
-      flatEntityId: destroyViewInput.id,
-      flatEntityMaps: recomputedExistingFlatViewMaps,
-    });
+    return flatViewFromDestroyInput;
   }
 }
