+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/services/workspace-migration-runner-v2.service.ts
@@ -74,6 +74,7 @@ export class WorkspaceMigrationRunnerV2Service {
 
       await queryRunner.commitTransaction();
 
+      console.log("RUNNER", flatEntityMapsToInvalidate);
       await this.flatEntityMapsCacheService.invalidateFlatEntityMaps({
         workspaceId,
         flatEntities: flatEntityMapsToInvalidate,
