+++ packages/twenty-server/test/integration/graphql/utils/graphql-test-assertions.util.ts
@@ -2,6 +2,8 @@ import {
   type BaseGraphQLError,
   type ErrorCode,
 } from 'src/engine/core-modules/graphql/utils/graphql-errors.util';
+import { jestExpectToBeDefined } from 'test/utils/expect-to-be-defined.util.test';
+import { extractRecordIdsAndDatesAsExpectAny } from 'test/utils/extract-record-ids-and-dates-as-expect-any';
 
 export interface GraphQLResponse<T extends Record<string, unknown>> {
   status: number;
@@ -32,14 +34,12 @@ export const assertGraphQLErrorResponse = <T extends Record<string, unknown>>(
   expectedErrorMessage?: string,
 ) => {
   expect(response.status).toBe(200);
-  expect(response.body.errors).toBeDefined();
+  jestExpectToBeDefined(response.body.errors);
   expect(response.body.errors).toHaveLength(1);
 
-  if (expectedErrorCode && response.body.errors) {
-    expect(response.body.errors[0].extensions?.code).toBe(expectedErrorCode);
-  }
-
-  if (expectedErrorMessage && response.body.errors) {
-    expect(response.body.errors[0].message).toBe(expectedErrorMessage);
-  }
+  const firstError = response.body.errors[0];
+  jestExpectToBeDefined(firstError);
+  expect(firstError).toMatchSnapshot(
+    extractRecordIdsAndDatesAsExpectAny(firstError),
+  );
 };
